backend/requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
cffi==1.14.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-import-export==2.4.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
